subtitlerBackend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiosignal==1.3.1
alembic==1.16.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asgiref==3.5.0
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==24.2.0
Automat==20.2.0
av==15.0.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.9.0
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.0
cliapp==1.20180812.1
click==8.1.7
cloud-init==25.1.2
cmake==3.25.0
colorama==0.4.4
coloredlogs==15.0.1
colorlog==6.9.0
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.3.2
cryptography==3.4.8
ctranslate2==4.4.0
cycler==0.12.1
Cython==0.29.36
datasets==3.1.0
dbus-python==1.2.18
dill==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
docopt==0.6.2
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.116.0
faster-whisper==1.1.1
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.2.10
fonttools==4.58.5
frozenlist==1.5.0
fsspec==2024.9.0
gguf==0.10.0
greenlet==3.2.3
h11==0.13.0
hf-xet==1.1.5
httplib2==0.20.2
huggingface-hub==0.33.2
humanfriendly==10.0
hyperlink==21.0.0
HyperPyYAML==1.2.2
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.5.1
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
julius==0.2.7
keyring==23.5.0
kiwisolver==1.4.8
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lightning==2.5.2
lightning-utilities==0.14.3
lit==15.0.7
llvmlite==0.43.0
Mako==1.3.10
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
netifaces==0.11.0
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==2.2.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.0
omegaconf==2.3.0
onnxruntime==1.22.0
openai-whisper @ git+https://github.com/openai/whisper.git@173ff7dd1d9fb1c4fddea0d41d704cfefeb8908c
optuna==4.4.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
primePy==1.3
propcache==0.2.0
protobuf==4.25.5
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==18.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.22
pycurl==7.44.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.2
pytorch-metric-learning==2.8.1
pytz==2024.2
PyYAML==5.4.1
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scikit-learn==1.7.0
scipy==1.15.3
SecretStorage==3.3.1
semver==3.0.4
sentencepiece==0.2.0
service-identity==18.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
systemd-python==234
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.21.2
tomli==2.2.1
torch==2.7.1+cu126
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.1+cu126
torchmetrics==1.7.4
tqdm==4.67.0
transformers==4.53.1
triton==3.3.1
ttystatus==0.38
Twisted==22.1.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.15.0
wadllib==1.3.6
Werkzeug==3.1.3
whisper==1.1.10
whisperx==3.4.2
wsproto==1.0.0
xxhash==3.5.0
yarl==1.17.1
zipp==1.0.0
zope.interface==5.4.0
